Extract the framing layer into a separate package
It is not really necessary because there will be only one layer but it will be convenient to have a layers package among other regular packages.
